Restructure user-service/ for clean architecture
