+++ user-service/internal/db/errors.go
@@ -1,4 +1,4 @@
-package postgres
+package db
 
 import (
 	"strings"
